backend/requirements.txt
# ========================================
# BroSync - Production Dependencies
# ========================================

# Django Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Django Channels (WebSocket support)
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Authentication (JWT)
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
psycopg2-binary>=2.9,<3.0
django-redis>=5.4,<6.0

# Celery (Async task processing)
celery>=5.4,<6.0
redis>=5.0,<6.0

# Docker SDK
docker>=7.0,<8.0

# Security
python-decouple>=3.8,<4.0
bleach>=6.1,<7.0
django-ratelimit>=4.1,<5.0

# Logging & Monitoring
structlog>=24.0,<25.0
django-structlog>=8.0,<9.0

# Utilities
python-dateutil>=2.9,<3.0
uuid>=1.30

# Development & Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.23,<1.0
factory-boy>=3.3,<4.0
coverage>=7.4,<8.0
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0

# ASGI server
uvicorn[standard]>=0.27,<1.0
gunicorn>=21.2,<23.0
